Bug 518664 - [lsp] DidChangeTextDocumentParams are sent back with incomplete parameters
When a document has been modified, the client must send a
notification back to the server to notify it of the change. If the
server supports incremental document synchronization, the range of
the document that has been modified must be included in the
notification. While TextDocumentContentChangeEvent also takes a
rangeLength parameter, due to the ambiguity of the protocol, we
should make sure both the range and the rangeLength is set properly
so that servers can use either piece of information to analyze the
changed document.

Signed-off-by: Remy Suen <remy.suen@gmail.com>