Bug 519425 - [lsp] Server doesn't get notified of changes if the change is at the beginning of the document
If a document is being modified at the beginning of the file, the
offset in the event will be zero. This will cause the boolean check
to be true and return early when monitoring for change events.
However, even if the user is modifying the beginning of the file, we
still want to notify the user so we should remove this zero offset
check.

Signed-off-by: Remy Suen <remy.suen@gmail.com>